Server/sustenido/python_socketio/requirements.txt
python-socketio
requests
websocket-client
numpy